refactor(canton): fold ledger helpers into client
